soilsync_backend/requirements.txt
Django==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.4
joblib==1.3.2
requests==2.31.0
python-dotenv==1.0.0
django-jazzmin==2.6.0
reportlab==4.0.7
